Move the new function to `fmn.database` to preserve the dependencies split
