+++ tests/database/model/test_destination.py
@@ -2,7 +2,7 @@
 #
 # SPDX-License-Identifier: MIT
 
-from datetime import datetime
+from datetime import datetime, timezone
 from email.utils import format_datetime
 from unittest import mock
 
@@ -10,7 +10,7 @@ import httpx
 import pytest
 
 from fmn.database import model
-from fmn.database.model.destination import Destination, get_extra
+from fmn.database.model.destination import Destination, get_extra, get_sent_datetime
 
 
 @pytest.fixture
@@ -145,3 +145,27 @@ async def test_get_extra_request_failure(make_mocked_message, respx_mocker):
 
     result = await get_extra(message)
     assert result == ""
+
+
+FAKE_NOW = datetime(2021, 1, 1, 0, 0, 0, tzinfo=timezone.utc)
+
+
+@pytest.mark.parametrize(
+    "sent_at,expected",
+    [
+        (None, FAKE_NOW),
+        ("2021-07-27T04:22:42Z", datetime(2021, 7, 27, 4, 22, 42, tzinfo=timezone.utc)),
+        ("2021-07-27T04:22:42JUNK", FAKE_NOW),
+    ],
+)
+def test_get_sent_datetime(make_mocked_message, mocker, sent_at, expected):
+    message = make_mocked_message(
+        topic="dummy",
+        body={"summary": "dummy summary"},
+    )
+    message._properties.headers["sent-at"] = sent_at
+
+    fake_datetime = mocker.patch("fmn.database.model.destination.datetime")
+    fake_datetime.now.return_value = FAKE_NOW
+    fake_datetime.fromisoformat = datetime.fromisoformat
+    assert get_sent_datetime(message) == expected
